llm-service/requirements.txt
annotated-types==0.7.0
certifi==2024.12.14
charset-normalizer==3.4.0
filelock==3.16.1
fsspec==2024.10.0
grpcio==1.59.0
grpcio-tools==1.59.0
huggingface-hub==0.17.3
idna==3.10
Jinja2==3.1.4
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.2.1
numpy==2.0.2
packaging==24.2
protobuf==4.24.4
pydantic==2.4.2
pydantic_core==2.10.1
python-dotenv==1.0.0
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.4.5
sympy==1.13.3
tokenizers==0.14.1
torch==2.1.0
tqdm==4.67.1
transformers==4.34.0
typing_extensions==4.12.2
urllib3==2.2.3
